Move ImageTarball from docker to image package
